refactor(apps): split catalog page modules
