+++ src/components/apps/AppsGoodToKnow.astro
@@ -1,0 +1,35 @@
+<section id="good-to-know" class="sect info-sect">
+  <div class="sect-head">
+    <span class="sect-icon">📋</span>
+    <div class="sect-titles"><h2>Good to Know</h2></div>
+  </div>
+
+  <div class="info-row">
+    <div class="info-card">
+      <h3>⬇ Installing Apps</h3>
+      <ul>
+        <li><strong>Mac App Store</strong> — Sandboxed, auto-updates</li>
+        <li><strong>Homebrew Cask</strong> — <code>brew install --cask &lt;app&gt;</code></li>
+        <li><strong>Direct download</strong> — From the developer's website</li>
+      </ul>
+      <div class="brew-box">
+        <code>brew install --cask raycast rectangle iterm2 visual-studio-code</code>
+      </div>
+    </div>
+    <div class="info-card">
+      <h3>🗑 Uninstalling Apps</h3>
+      <p>
+        Dragging an app to the Trash does not remove all its files.
+        Use <a href="https://freemacsoft.net/appcleaner/">AppCleaner</a> to fully
+        remove apps and their preferences, caches, and support&nbsp;files.
+      </p>
+    </div>
+    <div class="info-card">
+      <h3>📚 Resources</h3>
+      <p>
+        <a href="https://formulae.brew.sh/cask/">Homebrew Cask Formulae</a> — Search for
+        macOS applications available through Homebrew&nbsp;Cask.
+      </p>
+    </div>
+  </div>
+</section>
